made ui changes and performance enhancements

+++ server/src/v1/controllers/board.js
@@ -1,11 +1,10 @@
-const board = require('../models/board')
 const Board = require('../models/board')
 const Section = require('../models/section')
 const Task = require('../models/task')
 
 exports.create = async (req, res) => {
   try {
-    const boardsCount = await Board.find().countDocuments()
+    const boardsCount = await Board.countDocuments({ user: req.user._id })
     const board = await Board.create({
       user: req.user._id,
       position: boardsCount > 0 ? boardsCount : 0
@@ -79,13 +78,14 @@ exports.getAll = async (req, res) => {
 exports.updatePosition = async (req, res) => {
   const { boards } = req.body
   try {
-    for (const key in boards.reverse()) {
-      const board = boards[key]
-      await Board.findByIdAndUpdate(
-        board.id,
-        { $set: { position: key } }
-      )
-    }
+    const reversed = boards.reverse()
+    const ops = reversed.map((board, index) => ({
+      updateOne: {
+        filter: { _id: board.id },
+        update: { $set: { position: index } }
+      }
+    }))
+    await Board.bulkWrite(ops)
     res.status(200).json('updated')
   } catch (err) {
     res.status(500).json(err)
@@ -98,9 +98,23 @@ exports.getOne = async (req, res) => {
     const board = await Board.findOne({ user: req.user._id, _id: boardId })
     if (!board) return res.status(404).json('Board not found')
     const sections = await Section.find({ board: boardId })
+    const sectionIds = sections.map(s => s._id)
+
+    // Single query for all tasks across all sections instead of N+1
+    const allTasks = await Task.find({ section: { $in: sectionIds } })
+      .populate('section')
+      .sort('-position')
+
+    // Group tasks by section
+    const tasksBySection = {}
+    for (const task of allTasks) {
+      const sid = task.section._id.toString()
+      if (!tasksBySection[sid]) tasksBySection[sid] = []
+      tasksBySection[sid].push(task)
+    }
+
     for (const section of sections) {
-      const tasks = await Task.find({ section: section.id }).populate('section').sort('-position')
-      section._doc.tasks = tasks
+      section._doc.tasks = tasksBySection[section._id.toString()] || []
     }
     board._doc.sections = sections
     res.status(200).json(board)
@@ -124,17 +138,17 @@ exports.update = async (req, res) => {
         user: currentBoard.user,
         favourite: true,
         _id: { $ne: boardId }
-      }).sort('favouritePosition')
+      }).sort('favouritePosition').lean()
       if (favourite) {
         req.body.favouritePosition = favourites.length > 0 ? favourites.length : 0
-      } else {
-        for (const key in favourites) {
-          const element = favourites[key]
-          await Board.findByIdAndUpdate(
-            element.id,
-            { $set: { favouritePosition: key } }
-          )
-        }
+      } else if (favourites.length > 0) {
+        const ops = favourites.map((el, index) => ({
+          updateOne: {
+            filter: { _id: el._id },
+            update: { $set: { favouritePosition: index } }
+          }
+        }))
+        await Board.bulkWrite(ops)
       }
     }
 
@@ -150,7 +164,7 @@ exports.update = async (req, res) => {
 
 exports.getFavourites = async(req,res) => {
   try {
-    const favourites = await  Board.find({user:req.user._id,favourite:true }).sort('-favouritePosition')
+    const favourites = await Board.find({ user: req.user._id, favourite: true }).sort('-favouritePosition')
     res.status(200).json(favourites)
 
   } catch (err) {
@@ -162,13 +176,14 @@ exports.getFavourites = async(req,res) => {
 exports.updateFavouritePosition = async (req, res) => {
   const { boards } = req.body
   try {
-    for (const key in boards.reverse()) {
-      const board = boards[key]
-      await Board.findByIdAndUpdate(
-        board.id,
-        { $set: { favouritePosition: key } }
-      )
-    }
+    const reversed = boards.reverse()
+    const ops = reversed.map((board, index) => ({
+      updateOne: {
+        filter: { _id: board.id },
+        update: { $set: { favouritePosition: index } }
+      }
+    }))
+    await Board.bulkWrite(ops)
     res.status(200).json('updated')
   } catch (err) {
     res.status(500).json(err)
@@ -178,39 +193,45 @@ exports.updateFavouritePosition = async (req, res) => {
 exports.delete = async (req, res) => {
   const { boardId } = req.params
   try {
-    const sections = await Section.find({ board: boardId })
-    for (const section of sections) {
-      await Task.deleteMany({ section: section.id })
-    }
+    const currentBoard = await Board.findById(boardId)
+    if (!currentBoard) return res.status(404).json('Board not found')
+
+    // Delete all tasks for all sections of this board in one query
+    const sectionIds = (await Section.find({ board: boardId }).select('_id').lean()).map(s => s._id)
+    await Task.deleteMany({ section: { $in: sectionIds } })
     await Section.deleteMany({ board: boardId })
 
-    const currentBoard = await Board.findById(boardId)
-
+    // Reposition favourites if needed
     if (currentBoard.favourite) {
       const favourites = await Board.find({
         user: currentBoard.user,
         favourite: true,
         _id: { $ne: boardId }
-      }).sort('favouritePosition')
+      }).sort('favouritePosition').lean()
 
-      for (const key in favourites) {
-        const element = favourites[key]
-        await Board.findByIdAndUpdate(
-          element.id,
-          { $set: { favouritePosition: key } }
-        )
+      if (favourites.length > 0) {
+        const favOps = favourites.map((el, index) => ({
+          updateOne: {
+            filter: { _id: el._id },
+            update: { $set: { favouritePosition: index } }
+          }
+        }))
+        await Board.bulkWrite(favOps)
       }
     }
 
     await Board.deleteOne({ _id: boardId })
 
-    const boards = await Board.find().sort('position')
-    for (const key in boards) {
-      const board = boards[key]
-      await Board.findByIdAndUpdate(
-        board.id,
-        { $set: { position: key } }
-      )
+    // Reposition remaining boards for this user
+    const boards = await Board.find({ user: currentBoard.user }).sort('position').lean()
+    if (boards.length > 0) {
+      const posOps = boards.map((board, index) => ({
+        updateOne: {
+          filter: { _id: board._id },
+          update: { $set: { position: index } }
+        }
+      }))
+      await Board.bulkWrite(posOps)
     }
 
     res.status(200).json('deleted')
